+++ src/Checkin.test.js
@@ -135,14 +135,11 @@ jest.useFakeTimers();
 
 describe('CheckIn', () => {
   const removeEventListener = jest.fn((event, cb) => cb());
-  const addEventListener = jest.fn((event, cb) => cb({}));
-  const signal = jest.fn();
   const clear = jest.fn();
   createInactivityTimer.mockImplementation((time, cb) => {
     setTimeout(cb, time);
 
     return {
-      signal,
       clear,
     };
   });
@@ -157,7 +154,6 @@ describe('CheckIn', () => {
       removeEventListener: jest.fn(),
     });
     jest.spyOn(document, 'removeEventListener').mockImplementation(removeEventListener);
-    jest.spyOn(document, 'addEventListener').mockImplementation(addEventListener);
   });
 
   describe('Initial render', () => {
@@ -329,10 +325,6 @@ describe('CheckIn', () => {
         expect(clear).toHaveBeenCalled();
       });
 
-      it('should signal user activity', () => {
-        expect(signal).toHaveBeenCalled();
-      });
-
       it('should trigger createInactivityTimer with correct arguments', () => {
         const expectedArgs = [`${checkoutTimeoutDuration}m`, expect.any(Function)];
 
@@ -342,12 +334,6 @@ describe('CheckIn', () => {
       it('should remove listeners for keydown and mousedown events', () => {
         userActivityEvents.forEach((event) => {
           expect(removeEventListener).toHaveBeenCalledWith(event, expect.any(Function));
-        });
-      });
-
-      it('should add new listeners for keydown and mousedown events', () => {
-        userActivityEvents.forEach((event) => {
-          expect(addEventListener).toHaveBeenCalledWith(event, expect.any(Function));
         });
       });